requirements.txt
Flask
python-swiftclient
-e git+https://github.com/ggtracker/sc2reader.git@upstream#egg=sc2reader

